@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
     <PropertyGroup>
         <OutputType>Exe</OutputType>
-        <TargetFramework>net8.0</TargetFramework>
+        <TargetFramework>net9.0</TargetFramework>
         <LangVersion>latestmajor</LangVersion>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
@@ -34,12 +34,20 @@
     <ItemGroup>
       <PackageReference Include="Concentus" Version="2.2.2" />
       <PackageReference Include="Concentus.Oggfile" Version="1.0.6" />
-      <PackageReference Include="Discord.Net" Version="3.18.0" />
-      <PackageReference Include="Microsoft.Extensions.Configuration.UserSecrets" Version="8.0.0" />
-      <PackageReference Include="Microsoft.Extensions.Hosting" Version="8.0.0" />
-      <PackageReference Include="Microsoft.Extensions.Http" Version="8.0.0" />
-      <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="8.0.0" />
-      <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="8.0.0" />
+      <PackageReference Include="Microsoft.Extensions.Configuration.UserSecrets" Version="9.0.10" />
+      <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="9.0.10" />
+      <PackageReference Include="Microsoft.Extensions.Hosting" Version="9.0.10" />
+      <PackageReference Include="Microsoft.Extensions.Http" Version="9.0.10" />
+      <PackageReference Include="Microsoft.Extensions.Logging" Version="9.0.10" />
+      <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="9.0.10" />
+      <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="9.0.10" />
+      <PackageReference Include="NAudio" Version="2.2.1" />
+      <PackageReference Include="NetCord" Version="1.0.0-alpha.439" />
+      <PackageReference Include="NetCord.Hosting" Version="1.0.0-alpha.439" />
+      <PackageReference Include="NetCord.Services" Version="1.0.0-alpha.439" />
+      <PackageReference Include="NLog" Version="6.0.5" />
+      <PackageReference Include="NLog.Extensions.Logging" Version="6.0.5" />
+      <PackageReference Include="NLog.Web" Version="6.0.5" />
       <PackageReference Include="Sodium.Core" Version="1.4.0" />
       <PackageReference Include="Swashbuckle.AspNetCore.Annotations" Version="8.0.0" />
     </ItemGroup>
@@ -54,6 +62,9 @@
         <CopyToOutputDirectory>Always</CopyToOutputDirectory>
       </None>
       <None Update="assets\respawn_10m.wav">
+        <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+      </None>
+      <None Update="appsettings\nlog.config">
         <CopyToOutputDirectory>Always</CopyToOutputDirectory>
       </None>
     </ItemGroup>
