@@ -12,9 +12,18 @@ public static class ProxyEndpoints
         
         app.MapGet("/settings", async (IHttpClientFactory cf) =>
         {
-            var http = cf.CreateClient("worker");
-            var res = await http.GetFromJsonAsync<JsonElement>("/settings");
-            return Results.Ok(res);
+            try
+            {
+                var http = cf.CreateClient("worker");
+                var response = await http.GetAsync("/settings");
+                response.EnsureSuccessStatusCode();
+                var res = await response.Content.ReadFromJsonAsync<JsonElement>();
+                return Results.Ok(res);
+            }
+            catch (HttpRequestException ex)
+            {
+                return Results.Json(new { error = $"Worker error: {ex.Message}" }, statusCode: 503);
+            }
         });
 
         app.MapPost("/settings", async (IHttpClientFactory cf, HttpContext ctx) =>
@@ -54,18 +63,44 @@ public static class ProxyEndpoints
 
         app.MapGet("/settings/limits", async (IHttpClientFactory cf) =>
         {
-            var http = cf.CreateClient("worker");
-            var res = await http.GetFromJsonAsync<JsonElement>("/settings/limits");
-            return Results.Ok(res);
+            try
+            {
+                var http = cf.CreateClient("worker");
+                var response = await http.GetAsync("/settings/limits");
+                
+                if (response.StatusCode == System.Net.HttpStatusCode.NotFound)
+                {
+                    return Results.Json(new { error = "Endpoint not available", max_channels = 3 }, statusCode: 404);
+                    //return Results.Json(new { error = "Endpoint not available", limits = new object { } }, statusCode: 404);
+                }
+                
+                response.EnsureSuccessStatusCode();
+                var res = await response.Content.ReadFromJsonAsync<JsonElement>();
+                return Results.Ok(res);
+            }
+            catch
+            {
+                return Results.Json(new { error = "Worker unavailable", max_channels = 3 }, statusCode: 503);
+                //return Results.Json(new { error = "Worker unavailable", limits = new object { } }, statusCode: 503);
+            }
         });
 
         // === Respawn ===
         
         app.MapGet("/respawn/next", async (IHttpClientFactory cf) =>
         {
-            var http = cf.CreateClient("worker");
-            var res = await http.GetFromJsonAsync<JsonElement>("/respawn/next");
-            return Results.Ok(res);
+            try
+            {
+                var http = cf.CreateClient("worker");
+                var response = await http.GetAsync("/respawn/next");
+                response.EnsureSuccessStatusCode();
+                var res = await response.Content.ReadFromJsonAsync<JsonElement>();
+                return Results.Ok(res);
+            }
+            catch (HttpRequestException ex)
+            {
+                return Results.Json(new { error = $"Worker error: {ex.Message}" }, statusCode: 503);
+            }
         });
 
         app.MapPost("/respawn/sync", async (IHttpClientFactory cf) =>
@@ -98,11 +133,32 @@ public static class ProxyEndpoints
 
         // === Channels ===
         
-        app.MapGet("/channels/info", async (IHttpClientFactory cf) =>
+        app.MapPost("/channels/info", async (IHttpClientFactory cf, HttpContext ctx) =>
         {
-            var http = cf.CreateClient("worker");
-            var res = await http.GetFromJsonAsync<JsonElement>("/channels/info");
-            return Results.Ok(res);
+            try
+            {
+                var http = cf.CreateClient("worker");
+        
+                // Przekaż całe body z requestu do workera
+                using var reader = new StreamReader(ctx.Request.Body);
+                var jsonContent = await reader.ReadToEndAsync();
+        
+                var content = new StringContent(jsonContent, System.Text.Encoding.UTF8, "application/json");
+                var response = await http.PostAsync("/channels/info", content);
+        
+                if (response.StatusCode == System.Net.HttpStatusCode.MethodNotAllowed)
+                {
+                    return Results.Json(new { error = "Method not allowed" }, statusCode: 405);
+                }
+        
+                response.EnsureSuccessStatusCode();
+                var res = await response.Content.ReadFromJsonAsync<JsonElement>();
+                return Results.Ok(res);
+            }
+            catch
+            {
+                return Results.Json(new { error = "Worker unavailable" }, statusCode: 503);
+            }
         });
 
         app.MapGet("/channels/voice", async (IHttpClientFactory cf) =>
@@ -110,7 +166,16 @@ public static class ProxyEndpoints
             try
             {
                 var http = cf.CreateClient("worker");
-                var res = await http.GetFromJsonAsync<JsonElement>("/channels/voice");
+                var response = await http.GetAsync("/channels/voice");
+                
+                if (response.StatusCode == System.Net.HttpStatusCode.MethodNotAllowed)
+                {
+                    // Worker nie wspiera GET na tym endpoincie - zwróć pustą listę
+                    return Results.Json(new { channels = new object[] { } });
+                }
+                
+                response.EnsureSuccessStatusCode();
+                var res = await response.Content.ReadFromJsonAsync<JsonElement>();
                 return Results.Ok(res);
             }
             catch
