@@ -3,89 +3,85 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using Discord;
-using Discord.WebSocket;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
+using NetCord;
+using NetCord.Gateway;
+using NetCord.Rest;
+using NetCord.Services.ApplicationCommands;
 using Ogur.Sentinel.Worker.Discord.Modules;
 
 namespace Ogur.Sentinel.Worker.Discord;
 
-
 public sealed class CommandRegistrationService : BackgroundService
 {
-    private readonly DiscordSocketClient _client;
+    private readonly GatewayClient _client;
     private readonly ILogger<CommandRegistrationService> _logger;
-    private readonly RespawnModule _respawn;
-    private readonly LeaveModule _leave;
-    private readonly AdminBreakModule _admin;
+    private readonly ApplicationCommandService<SlashCommandContext> _commandService;
 
     public CommandRegistrationService(
-        DiscordSocketClient client,
-        RespawnModule respawn,
-        LeaveModule leave,
-        AdminBreakModule admin,
+        GatewayClient client,
+        ApplicationCommandService<SlashCommandContext> commandService,
         ILogger<CommandRegistrationService> logger)
     {
         _client = client;
+        _commandService = commandService;
         _logger = logger;
-        _respawn = respawn;
-        _leave = leave;
-        _admin = admin;
     }
 
     protected override async Task ExecuteAsync(CancellationToken stoppingToken)
     {
-        _client.SlashCommandExecuted += OnSlash;
+        _client.InteractionCreate += OnInteraction;
 
-        await _client.WaitForReadyAsync(stoppingToken);
+        await WaitForReadyAsync(stoppingToken);
 
-        var commands = new List<SlashCommandBuilder>
-        {
-            _respawn.Build(),
-            _leave.Build(),
-            _admin.Build()
-        };
-
-        foreach (var g in _client.Guilds)
-        {
-            try
-            {
-                await g.BulkOverwriteApplicationCommandAsync(commands.Select(c => c.Build()).ToArray());
-                _logger.LogInformation("Slash commands registered for guild {Guild}", g.Name);
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex, "Failed to register slash commands for guild {Guild}", g.Name);
-            }
-        }
+        _logger.LogInformation("Command registration service started. Commands will be handled via ApplicationCommandService.");
     }
 
-    private Task OnSlash(SocketSlashCommand cmd) =>
-        cmd.Data.Name switch
-        {
-            "respawn" => _respawn.Handle(cmd),
-            "leave"   => _leave.Handle(cmd),
-            "admin"   => _admin.Handle(cmd),
-            _ => cmd.RespondAsync("Unknown command.", ephemeral: true)
-        };
-}
-
-file static class DiscordClientExtensions
-{
-    public static async Task WaitForReadyAsync(this DiscordSocketClient client, CancellationToken ct)
+    private ValueTask OnInteraction(Interaction interaction)
     {
-        if (client.LoginState == LoginState.LoggedIn && client.ConnectionState == ConnectionState.Connected && client.CurrentUser != null)
+        if (interaction is SlashCommandInteraction slashCommand)
+        {
+            var context = new SlashCommandContext(slashCommand, _client);
+            _ = Task.Run(async () =>
+            {
+                try
+                {
+                    await _commandService.ExecuteAsync(context);
+                }
+                catch (Exception ex)
+                {
+                    _logger.LogError(ex, "Error executing slash command {CommandName}", slashCommand.Data.Name);
+                }
+            });
+        }
+
+        return ValueTask.CompletedTask;
+    }
+
+    private async Task WaitForReadyAsync(CancellationToken ct)
+    {
+        // Sprawdź czy client już jest gotowy - najprostsze rozwiązanie
+        if (_client.Cache?.Guilds != null && _client.Cache.Guilds.Count > 0)
             return;
 
         var tcs = new TaskCompletionSource();
-        Task Handler()
+        
+        ValueTask Handler(ReadyEventArgs args)
         {
             tcs.TrySetResult();
-            return Task.CompletedTask;
+            return ValueTask.CompletedTask;
         }
-        client.Ready += Handler;
-        try { await Task.WhenAny(tcs.Task, Task.Delay(TimeSpan.FromSeconds(30), ct)); }
-        finally { client.Ready -= Handler; }
+
+        _client.Ready += Handler;
+        
+        try
+        {
+            await Task.WhenAny(tcs.Task, Task.Delay(TimeSpan.FromSeconds(30), ct));
+        }
+        finally
+        {
+            _client.Ready -= Handler;
+        }
     }
 }
