@@ -1,8 +1,9 @@
 ﻿using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options; 
 using Ogur.Sentinel.Abstractions.Options;
-using Discord;
-using Discord.WebSocket;
+using NetCord;
+using NetCord.Rest;
+using NetCord.Services.ApplicationCommands;
 using Ogur.Sentinel.Abstractions.Respawn;
 using Ogur.Sentinel.Core.Respawn;
 using Ogur.Sentinel.Worker.Services;
@@ -11,16 +12,22 @@ using System.Threading.Tasks;
 
 namespace Ogur.Sentinel.Worker.Discord.Modules;
 
-public sealed class RespawnModule
+[SlashCommand("respawn", "Respawn management")]
+public sealed class RespawnModule : ApplicationCommandModule<SlashCommandContext>
 {
     private readonly RespawnState _state;
     private readonly SettingsStore _store;
     private readonly RespawnSchedulerService _scheduler;
-    private readonly VoiceService2 _voice;
+    private readonly VoiceService3 _voice;
     private readonly RespawnOptions _respawn; 
     private readonly ILogger<RespawnModule> _logger;
 
-    public RespawnModule(RespawnState state, SettingsStore store, RespawnSchedulerService scheduler, VoiceService2 voice, IOptions<RespawnOptions> opts,
+    public RespawnModule(
+        RespawnState state, 
+        SettingsStore store, 
+        RespawnSchedulerService scheduler, 
+        VoiceService3 voice, 
+        IOptions<RespawnOptions> opts,
         ILogger<RespawnModule> logger)
     {
         _state = state;
@@ -31,171 +38,172 @@ public sealed class RespawnModule
         _logger = logger;
     }
 
-    public SlashCommandBuilder Build()
+    [SubSlashCommand("on", "Enable respawn reminders")]
+    public async Task OnAsync(SlashCommandContext context)
     {
-        var b = new SlashCommandBuilder()
-            .WithName("respawn")
-            .WithDescription("Respawn management");
-
-        b.AddOption(new SlashCommandOptionBuilder()
-            .WithName("on")
-            .WithDescription("Enable respawn reminders")
-            .WithType(ApplicationCommandOptionType.SubCommand));
-
-        b.AddOption(new SlashCommandOptionBuilder()
-            .WithName("off")
-            .WithDescription("Disable respawn reminders")
-            .WithType(ApplicationCommandOptionType.SubCommand));
-
-        b.AddOption(new SlashCommandOptionBuilder()
-            .WithName("set-base")
-            .WithDescription("Set base time (format HH:MM:SS)")
-            .AddOption("time", ApplicationCommandOptionType.String, "Base time e.g. 01:10:30", isRequired: true)  // <-- przykład
-            .WithType(ApplicationCommandOptionType.SubCommand));
-
-        b.AddOption(new SlashCommandOptionBuilder()
-            .WithName("set-lead")
-            .WithDescription("Set lead seconds")
-            .AddOption("seconds", ApplicationCommandOptionType.Integer, "Lead seconds", isRequired: true)
-            .WithType(ApplicationCommandOptionType.SubCommand));
-
-        b.AddOption(new SlashCommandOptionBuilder()
-            .WithName("add-channel")
-            .WithDescription("Add voice channel id for join")
-            .AddOption("channel_id", ApplicationCommandOptionType.String, "Channel Id (ulong)", isRequired: true)
-            .WithType(ApplicationCommandOptionType.SubCommand));
-
-        b.AddOption(new SlashCommandOptionBuilder()
-            .WithName("remove-channel")
-            .WithDescription("Remove voice channel id")
-            .AddOption("channel_id", ApplicationCommandOptionType.String, "Channel Id (ulong)", isRequired: true)
-            .WithType(ApplicationCommandOptionType.SubCommand));
-
-        b.AddOption(new SlashCommandOptionBuilder()
-            .WithName("test-voice")
-            .WithDescription("Join channel and play test sound")
-            .WithType(ApplicationCommandOptionType.SubCommand)
-            .AddOption("sound", ApplicationCommandOptionType.String, "Which sound to play (10m or 2h)", isRequired: true,
-                choices: new []
-                {
-                    new ApplicationCommandOptionChoiceProperties { Name = "10m respawn", Value = "10m" },
-                    new ApplicationCommandOptionChoiceProperties { Name = "2h respawn", Value = "2h" }
-                })
-            .AddOption("channel", ApplicationCommandOptionType.String, "Voice channel ID (leave empty for first configured)", isRequired: false));
-
-        return b;
+        _state.Enabled10m = true;
+        _state.Enabled2h = true;
+        await context.Interaction.SendResponseAsync(InteractionCallback.Message(new InteractionMessageProperties
+        {
+            Content = "Respawn enabled (10m & 2h).",
+            Flags = MessageFlags.Ephemeral
+        }));
     }
 
-    public async Task Handle(SocketSlashCommand cmd)
+    [SubSlashCommand("off", "Disable respawn reminders")]
+    public async Task OffAsync(SlashCommandContext context)
     {
-        var sub = cmd.Data.Options.FirstOrDefault()?.Name;
-        switch (sub)
+        _state.Enabled10m = false;
+        _state.Enabled2h = false;
+        await context.Interaction.SendResponseAsync(InteractionCallback.Message(new InteractionMessageProperties
         {
-            case "on":
-                _state.Enabled10m = true;
-                _state.Enabled2h = true;
-                await cmd.RespondAsync("Respawn enabled (10m & 2h).", ephemeral: true);
-                break;
+            Content = "Respawn disabled.",
+            Flags = MessageFlags.Ephemeral
+        }));
+    }
 
-            case "off":
-                _state.Enabled10m = false;
-                _state.Enabled2h = false;
-                await cmd.RespondAsync("Respawn disabled.", ephemeral: true);
-                break;
+    [SubSlashCommand("set-base", "Set base time (format HH:MM:SS)")]
+    public async Task SetBaseAsync(
+        SlashCommandContext context,
+        [SlashCommandParameter(Name = "time", Description = "Base time e.g. 01:10:30")] string time)
+    {
+        SchedulingMath.ParseHhmm(time); // validate
+        _state.BaseHhmm = time;
+        await _store.SaveAsync(_state.ToPersisted());
+        await context.Interaction.SendResponseAsync(InteractionCallback.Message(new InteractionMessageProperties
+        {
+            Content = $"Base set to `{time}`.",
+            Flags = MessageFlags.Ephemeral
+        }));
+    }
 
-            case "set-base":
-                var t = (string)cmd.Data.Options.First().Options.First().Value!;
-                SchedulingMath.ParseHhmm(t); // validate
-                _state.BaseHhmm = t;
-                await _store.SaveAsync(_state.ToPersisted());
-                await cmd.RespondAsync($"Base set to `{t}`.", ephemeral: true);
-                break;
+    [SubSlashCommand("set-lead", "Set lead seconds")]
+    public async Task SetLeadAsync(
+        SlashCommandContext context,
+        [SlashCommandParameter(Name = "seconds", Description = "Lead seconds")] int seconds)
+    {
+        _state.LeadSeconds = Math.Max(0, seconds);
+        await _store.SaveAsync(_state.ToPersisted());
+        await context.Interaction.SendResponseAsync(InteractionCallback.Message(new InteractionMessageProperties
+        {
+            Content = $"Lead set to `{_state.LeadSeconds}s`.",
+            Flags = MessageFlags.Ephemeral
+        }));
+    }
 
-            case "set-lead":
-                var s = Convert.ToInt32(cmd.Data.Options.First().Options.First().Value!);
-                _state.LeadSeconds = Math.Max(0, s);
-                await _store.SaveAsync(_state.ToPersisted());
-                await cmd.RespondAsync($"Lead set to `{_state.LeadSeconds}s`.", ephemeral: true);
-                break;
-
-            case "add-channel":
+    [SubSlashCommand("add-channel", "Add voice channel id for join")]
+    public async Task AddChannelAsync(
+        SlashCommandContext context,
+        [SlashCommandParameter(Name = "channel_id", Description = "Channel Id (ulong)")] string channelIdStr)
+    {
+        if (ulong.TryParse(channelIdStr, out var id))
+        {
+            var newList = _state.Channels.ToList();
+            if (!newList.Contains(id)) newList.Add(id);
+            _state.SetChannels(newList);
+            await _store.SaveAsync(_state.ToPersisted());
+            await context.Interaction.SendResponseAsync(InteractionCallback.Message(new InteractionMessageProperties
             {
-                var str = (string)cmd.Data.Options.First().Options.First().Value!;
-                if (ulong.TryParse(str, out var id))
-                {
-                    var newList = _state.Channels.ToList();
-                    if (!newList.Contains(id)) newList.Add(id);
-                    _state.SetChannels(newList);
-                    await _store.SaveAsync(_state.ToPersisted());
-                    await cmd.RespondAsync($"Added channel `{id}`.", ephemeral: true);
-                }
-                else await cmd.RespondAsync("Invalid channel id.", ephemeral: true);
-
-                break;
-            }
-
-            case "remove-channel":
+                Content = $"Added channel `{id}`.",
+                Flags = MessageFlags.Ephemeral
+            }));
+        }
+        else
+        {
+            await context.Interaction.SendResponseAsync(InteractionCallback.Message(new InteractionMessageProperties
             {
-                var str = (string)cmd.Data.Options.First().Options.First().Value!;
-                if (ulong.TryParse(str, out var id))
-                {
-                    if (_state.RemoveChannel(id))
-                    {
-                        await _store.SaveAsync(_state.ToPersisted());
-                        await cmd.RespondAsync($"Removed `{id}`.", ephemeral: true);
-                    }
-                    else await cmd.RespondAsync("Not found.", ephemeral: true);
-                }
-                else await cmd.RespondAsync("Invalid channel id.", ephemeral: true);
-
-                break;
-            }
-
-            case "test-voice":
-            {
-                var subCommand = cmd.Data.Options.First();
-                var soundType = subCommand.Options.FirstOrDefault(o => o.Name == "sound")?.Value?.ToString() ?? "10m";
-                var channelIdStr = subCommand.Options.FirstOrDefault(o => o.Name == "channel")?.Value?.ToString();
-    
-                ulong channelId;
-                
-                if (!string.IsNullOrEmpty(channelIdStr) && ulong.TryParse(channelIdStr, out var parsed))
-                {
-                    channelId = parsed;
-                }
-  
-                else
-                {
-                    channelId = _state.Channels.FirstOrDefault();
-                    if (channelId == 0)
-                    {
-                        await cmd.RespondAsync("Brak skonfigurowanych kanałów. Użyj `respawn add-channel` lub podaj ID kanału.", ephemeral: true);
-                        return;
-                    }
-                }
-                
-                var path = soundType == "2h" ? _respawn.Sound2h : _respawn.Sound10m;
-                path ??= soundType == "2h" ? "/assets/respawn_2h.wav" : "/assets/respawn_10m.wav";
-    
-                if (!File.Exists(path))
-                {
-                    await cmd.RespondAsync($"Plik audio nie istnieje: `{path}`", ephemeral: true);
-                    return;
-                }
-                
-                var repeatPlays = _state.RepeatPlays > 0 ? _state.RepeatPlays : 1;
-                var repeatGapMs = _state.RepeatGapMs;  // ← Bez fallbacku
-
-                await cmd.RespondAsync($"OK, próbuję wejść na kanał `{channelId}` i zagrać ({soundType}, {repeatPlays}x, gap={repeatGapMs}ms).", ephemeral: true);
-                _ = _voice.JoinAndPlayAsync(channelId, path, repeatPlays, repeatGapMs, CancellationToken.None);
-                break;
-            }
-
-            
-
-            default:
-                await cmd.RespondAsync("Unknown subcommand.", ephemeral: true);
-                break;
+                Content = "Invalid channel id.",
+                Flags = MessageFlags.Ephemeral
+            }));
         }
     }
+
+    [SubSlashCommand("remove-channel", "Remove voice channel id")]
+    public async Task RemoveChannelAsync(
+        SlashCommandContext context,
+        [SlashCommandParameter(Name = "channel_id", Description = "Channel Id (ulong)")] string channelIdStr)
+    {
+        if (ulong.TryParse(channelIdStr, out var id))
+        {
+            if (_state.RemoveChannel(id))
+            {
+                await _store.SaveAsync(_state.ToPersisted());
+                await context.Interaction.SendResponseAsync(InteractionCallback.Message(new InteractionMessageProperties
+                {
+                    Content = $"Removed `{id}`.",
+                    Flags = MessageFlags.Ephemeral
+                }));
+            }
+            else
+            {
+                await context.Interaction.SendResponseAsync(InteractionCallback.Message(new InteractionMessageProperties
+                {
+                    Content = "Not found.",
+                    Flags = MessageFlags.Ephemeral
+                }));
+            }
+        }
+        else
+        {
+            await context.Interaction.SendResponseAsync(InteractionCallback.Message(new InteractionMessageProperties
+            {
+                Content = "Invalid channel id.",
+                Flags = MessageFlags.Ephemeral
+            }));
+        }
+    }
+
+    [SubSlashCommand("test-voice", "Join channel and play test sound")]
+public async Task TestVoiceAsync(
+    SlashCommandContext context,
+    [SlashCommandParameter(Name = "sound", Description = "Which sound to play (10m or 2h)")]
+    string sound,
+    [SlashCommandParameter(Name = "channel", Description = "Voice channel ID (leave empty for first configured)")] 
+    string? channelIdStr = null)
+{
+    ulong channelId;
+    
+    if (!string.IsNullOrEmpty(channelIdStr) && ulong.TryParse(channelIdStr, out var parsed))
+    {
+        channelId = parsed;
+    }
+    else
+    {
+        channelId = _state.Channels.FirstOrDefault();
+        if (channelId == 0)
+        {
+            await context.Interaction.SendResponseAsync(InteractionCallback.Message(new InteractionMessageProperties
+            {
+                Content = "Brak skonfigurowanych kanałów. Użyj `respawn add-channel` lub podaj ID kanału.",
+                Flags = MessageFlags.Ephemeral
+            }));
+            return;
+        }
+    }
+    
+    var path = sound == "2h" ? _respawn.Sound2h : _respawn.Sound10m;
+    path ??= sound == "2h" ? "/assets/respawn_2h.wav" : "/assets/respawn_10m.wav";
+
+    if (!File.Exists(path))
+    {
+        await context.Interaction.SendResponseAsync(InteractionCallback.Message(new InteractionMessageProperties
+        {
+            Content = $"Plik audio nie istnieje: `{path}`",
+            Flags = MessageFlags.Ephemeral
+        }));
+        return;
+    }
+    
+    // ✅ Użyj odpowiednich settings dla danego dźwięku
+    var repeatPlays = sound == "2h" ? _state.RepeatPlays2h : _state.RepeatPlays10m;
+    var repeatGapMs = sound == "2h" ? _state.RepeatGapMs2h : _state.RepeatGapMs10m;
+
+    await context.Interaction.SendResponseAsync(InteractionCallback.Message(new InteractionMessageProperties
+    {
+        Content = $"OK, próbuję wejść na kanał `{channelId}` i zagrać ({sound}, {repeatPlays}x, gap={repeatGapMs}ms).",
+        Flags = MessageFlags.Ephemeral
+    }));
+    
+    _ = _voice.JoinAndPlayAsync(channelId, path, repeatPlays, repeatGapMs, CancellationToken.None);
+}
 }
